Move showToast from utilTemplateBuilder into utilCommon
