+++ src/lwc/geListView/geListView.js
@@ -5,13 +5,13 @@ import {
     dispatch,
     handleError,
     generateId,
-    showToast,
 } from 'c/utilTemplateBuilder'
 import {
     format,
     deepClone,
     isNotEmpty,
     hasNestedProperty,
+    showToast,
 } from 'c/utilCommon'
 import { fireEvent } from 'c/pubsubNoPageRef'
 import LibsMoment from 'c/libsMoment';
