+++ src/lwc/rd2EntryForm/rd2EntryForm.js-meta.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <LightningComponentBundle xmlns="http://soap.sforce.com/2006/04/metadata">
     <apiVersion>48.0</apiVersion>
+    <description>Enhanced Recurring Donations New and Edit Parent Form</description>
     <isExposed>false</isExposed>
 </LightningComponentBundle>
