+++ src/lwc/utilTemplateBuilder/utilTemplateBuilder.js
@@ -1,6 +1,6 @@
 /* eslint-disable @lwc/lwc/no-async-operation */
 import { ShowToastEvent } from 'lightning/platformShowToastEvent'
-import { isEmpty, isNotEmpty, deepClone } from 'c/utilCommon';
+import { isEmpty, isNotEmpty, deepClone, showToast } from 'c/utilCommon';
 
 // Import schema for additionally required fields for the template batch header
 import DI_BATCH_NAME_FIELD_INFO from '@salesforce/schema/DataImportBatch__c.Name';
@@ -71,7 +71,7 @@ import getGiftEntrySettings from
 const CONTACT1 = 'Contact1';
 const ACCOUNT1 = 'Account1';
 
-// cache custom labels 
+// cache custom labels
 const CUSTOM_LABELS = GeLabelService.CUSTOM_LABELS;
 
 const ADVANCED_MAPPING = 'Data Import Field Mapping';
@@ -262,28 +262,6 @@ const dispatch = (context, name, detail, bubbles = false, composed = false) => {
 }
 
 /*******************************************************************************
-* @description Creates and dispatches a ShowToastEvent
-*
-* @param {string} title: Title of the toast, displayed as a heading.
-* @param {string} message: Message of the toast. It can contain placeholders in
-* the form of {0} ... {N}. The placeholders are replaced with the links from
-* messageData param
-* @param {string} mode: Mode of the toast
-* @param {array} messageData: List of values that replace the {index} placeholders
-* in the message param
-*/
-const showToast = (title, message, variant, mode, messageData) => {
-    const event = new ShowToastEvent({
-        title: title,
-        message: message,
-        variant: variant,
-        mode: mode,
-        messageData: messageData
-    });
-    dispatchEvent(event);
-}
-
-/*******************************************************************************
 * @description Creates and dispatches an error toast.
 *
 * @param {object} error: Event holding error details
@@ -363,7 +341,7 @@ const getRecordFieldNames = (formTemplate, fieldMappings, apiName) => {
                             let fieldName = fieldMappings[fieldMappingDevName].Target_Field_API_Name;
                             fieldNames.push(`${objectName}.${fieldName}`);
                         }
-                    }                
+                    }
                 }
             }
 
@@ -395,10 +373,10 @@ const checkPermissionErrors = (formTemplate) => {
 
     if (template.permissionErrorType === CRUD_ERROR_TYPE) {
         errorObject.errorTitle = CUSTOM_LABELS.geErrorObjectCRUDHeader;
-        errorObject.errorMessage = GeLabelService.format(CUSTOM_LABELS.geErrorObjectCRUDBody, permissionErrors);   
+        errorObject.errorMessage = GeLabelService.format(CUSTOM_LABELS.geErrorObjectCRUDBody, permissionErrors);
     } else if (template.permissionErrorType === FLS_ERROR_TYPE) {
         errorObject.errorTitle = CUSTOM_LABELS.geErrorFLSHeader;
-        errorObject.errorMessage = GeLabelService.format(CUSTOM_LABELS.geErrorFLSBody, permissionErrors); 
+        errorObject.errorMessage = GeLabelService.format(CUSTOM_LABELS.geErrorFLSBody, permissionErrors);
     }
 
     return errorObject;
@@ -424,7 +402,7 @@ const setRecordValuesOnTemplate = (templateSections, fieldMappings, record) => {
 
     sections.forEach(section => {
         const elements = section.elements;
-  
+
         elements.forEach(element => {
             if (element.elementType === 'field') {
                 // set an empty default value
@@ -514,7 +492,6 @@ export {
     DONATION_DONOR_FIELDS,
     DONATION_DONOR,
     dispatch,
-    showToast,
     handleError,
     generateId,
     inputTypeByDescribeType,
